moving files and renaming
