+++ src/burwell-sources/archive_/content.ts
@@ -1,6 +1,6 @@
-import IndexMapContainer from "./index-map";
+import IndexMapContainer from "../components/map";
 import h from "@macrostrat/hyper";
-import { InfoDrawer } from "./info-drawer";
+import { InfoDrawer } from "../components/info-drawer";
 
 const Content = () => {
   return h("div.full-height", [h(IndexMapContainer), h(InfoDrawer)]);
