--- src/burwell-sources/components/map-pieces/index.ts
@@ -1,2 +1,0 @@
-export * from "./initialize-map";
-export * from "./map-sources";
